Rename module ledgersimpleaddresskeyringentry -> ledgersimpleaddresswallet
