+++ packages/iov-ledger-bns/src/index.ts
@@ -1,2 +1,2 @@
-export { LedgerSimpleAddressWallet } from "./ledgersimpleaddresskeyringentry";
+export { LedgerSimpleAddressWallet } from "./ledgersimpleaddresswallet";
 export { LedgerState } from "./statetracker";
