+++ packages/iov-ledger-bns/types/index.d.ts
@@ -1,2 +1,2 @@
-export { LedgerSimpleAddressWallet } from "./ledgersimpleaddresskeyringentry";
+export { LedgerSimpleAddressWallet } from "./ledgersimpleaddresswallet";
 export { LedgerState } from "./statetracker";
